Refactor rewards
