+++ package.json
@@ -41,7 +41,7 @@
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
     "react-native-elements": "^1.2.6",
     "react-native-gesture-handler": "~1.3.0",
-    "react-native-maps": "^0.26.1",
+    "react-native-maps": "~0.25.0",
     "react-native-phone-input": "^0.2.4",
     "react-native-reanimated": "~1.2.0",
     "react-native-screens": "^1.0.0-alpha.23",
