+++ screens/rewards/PointsHistoryScreen.js
@@ -2,8 +2,8 @@ import * as Font from 'expo-font';
 import React from 'react';
 import { AsyncStorage, ScrollView, Text, View } from 'react-native';
 
-import Transactions from '../components/Transactions';
-import getCustomerTransactions from './historyHelpers';
+import Transactions from '../../components/Transactions';
+import { getCustomerTransactions } from './rewardsHelpers';
 
 export default class PointsHistoryScreen extends React.Component {
   constructor(props) {
@@ -22,7 +22,7 @@ export default class PointsHistoryScreen extends React.Component {
     });
 
     await Font.loadAsync({
-      Poppins: require('../assets/fonts/Poppins-Regular.ttf')
+      Poppins: require('../../assets/fonts/Poppins-Regular.ttf')
     });
   }
